pyproject.toml
[project]
name = "visualmind"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aisuite[all]",
    "networkx",
    "pyvis",
    "python-dotenv",
    "plotly>=6.3.1",
    "ipykernel==6.30.1",
    "numpy>=2.3.4",
    "nbformat>=5.10.4",
]
